feature(GoogleAnalytics): Add support for AMP-ANALYTICS if the site have tracking system
